Updated the program setup for a more divided interface
